Fixed type of fields

+++ io-hotmoka-helpers/src/main/java/io/hotmoka/helpers/internal/ManifestHelperImpl.java
@@ -154,13 +154,13 @@ public class ManifestHelperImpl implements ManifestHelper {
 			String chainId = getChainId();
 			builder.append("   ├─ chainId: ").append(chainId).append("\n");
 
-			long maxErrorLength = ((LongValue) node.runInstanceMethodCallTransaction(TransactionRequests.instanceViewMethodCall
+			int maxErrorLength = ((IntValue) node.runInstanceMethodCallTransaction(TransactionRequests.instanceViewMethodCall
 				(manifest, _100_000, takamakaCode, MethodSignatures.GET_MAX_ERROR_LENGTH, manifest))
 				.orElseThrow(() -> new NodeException(MethodSignatures.GET_MAX_ERROR_LENGTH + " should not return void"))).getValue();
 
 			builder.append("   ├─ maxErrorLength: ").append(maxErrorLength).append("\n");
 
-			long maxDependencies = ((LongValue) node.runInstanceMethodCallTransaction(TransactionRequests.instanceViewMethodCall
+			int maxDependencies = ((IntValue) node.runInstanceMethodCallTransaction(TransactionRequests.instanceViewMethodCall
 				(manifest, _100_000, takamakaCode, MethodSignatures.GET_MAX_DEPENDENCIES, manifest))
 				.orElseThrow(() -> new NodeException(MethodSignatures.GET_MAX_DEPENDENCIES + " should not return void"))).getValue();
